HyperspaceWebService/requirements.txt
azure==4.0.0
pandas==0.23.0
pandasql=0.7.3
requests2==2.16.0
PyYAML==3.13
Flask==1.0.2
mysqlclient==1.3.13
